src/user/infra/db/in-memory/user-in-memory.repository.spec.ts
import { User } from '../../../domain';
import { UserInMemoryRepository } from './user-in-memory.repository';

describe('UserInMemoryRepository', () => {
  let repository: UserInMemoryRepository;

  beforeEach(() => (repository = new UserInMemoryRepository()));

  it('should no filter items when filter object is null', async () => {
    const items = [User.fake().aUser().build(), User.fake().aUser().build()];
    const filterSpy = jest.spyOn(items, 'filter' as any);

    const itemsFiltered = await repository['applyFilter'](items, null);
    expect(filterSpy).not.toHaveBeenCalled();
    expect(itemsFiltered).toStrictEqual(items);
  });

  it('should sort by createdAt when sort param is null', async () => {
    const items = [
      User.fake().aUser().withCreatedAt(new Date()).build(),
      User.fake()
        .aUser()
        .withCreatedAt(new Date(new Date().getTime() + 1))
        .build(),
      User.fake()
        .aUser()
        .withCreatedAt(new Date(new Date().getTime() + 2))
        .build(),
    ];

    const itemsSorted = await repository['applySort'](items, null, null);
    expect(itemsSorted).toStrictEqual([items[2], items[1], items[0]]);
  });

  it('should sort by user name', async () => {
    const items = [
      User.fake().aUser().withName('c').build(),
      User.fake().aUser().withName('b').build(),
      User.fake().aUser().withName('a').build(),
    ];

    let itemsSorted = await repository['applySort'](items, 'name', 'asc');
    expect(itemsSorted).toStrictEqual([items[2], items[1], items[0]]);

    itemsSorted = await repository['applySort'](items, 'name', 'desc');
    expect(itemsSorted).toStrictEqual([items[0], items[1], items[2]]);
  });

  it('should filter items by user name', async () => {
    const faker = User.fake().aUser();
    const items = [
      faker.withName('test').build(),
      faker.withName('TEST').build(),
      faker.withName('fake').build(),
    ];
    const filterSpy = jest.spyOn(items, 'filter' as any);

    const itemsFiltered = await repository['applyFilter'](items, {
      name: 'TEST',
    });
    expect(filterSpy).toHaveBeenCalledTimes(1);
    expect(itemsFiltered).toStrictEqual([items[0], items[1]]);
  });

  it('should filter items by email', async () => {
    const faker = User.fake().aUser();
    const items = [
      faker.withEmail('test').build(),
      faker.withEmail('TEST').build(),
      faker.withEmail('fake').build(),
    ];
    const filterSpy = jest.spyOn(items, 'filter' as any);

    const itemsFiltered = await repository['applyFilter'](items, {
      email: 'TEST',
    });
    expect(filterSpy).toHaveBeenCalledTimes(1);
    expect(itemsFiltered).toStrictEqual([items[0], items[1]]);
  });
});
